Moved signed status properties & methods to reusable class trait
